# Geometry example in the examples folder is broken #2708