Since some update the window stopped receiving updates constantly, but only when the mouse entered and exited the window. 

Example `Geometry` broke - there was no reaction to mouse movement. Now the original correct behavior has been restored.