@ArtHome12 2025-03-07
# Benchmark 
## Before
- default size 120 FPS, 25TPS, 
- with size in whole world 3-4 FPS, 25 TPS.
- full screen (with filter, the world fits 10 times), refresh every 1.5 seconds. Full CPU utilization.

## With bitmap unoptimized
- default size 120 FPS, 25TPS, 
- with size in whole world 6 FPS, 25 TPS.
- full screen refresh every 1-2 FPS, 25-42 TPS.