@hecrj 2025-11-21
Thanks!

But the proper fix is #3020.